+++ src/auth/session/storage/mysql.ts
@@ -95,13 +95,13 @@ export class MySQLSessionStorage implements SessionStorage {
 
   public async findSessionsByShop(shop: string): Promise<SessionInterface[]> {
     await this.ready;
-    sanitizeShop(shop, true);
+    const cleanShop = sanitizeShop(shop, true)!;
 
     const query = `
       SELECT * FROM ${this.options.sessionTableName}
       WHERE shop = ?;
     `;
-    const [rows] = await this.query(query, [shop]);
+    const [rows] = await this.query(query, [cleanShop]);
     if (!Array.isArray(rows) || rows?.length === 0) return [];
 
     const results: SessionInterface[] = rows.map((row) => {